requirements.txt
# ============================================================
# Anti-Gravity Deepfake Detection System — Requirements
# Python 3.10+
# ============================================================

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Model Architectures
timm>=0.9.12               # EfficientNet-B4, ViT models
facenet-pytorch>=2.5.3     # MTCNN face detection

# Computer Vision
opencv-python>=4.8.1
Pillow>=10.1.0
albumentations>=1.3.1
scikit-image>=0.22.0

# Data Science
numpy>=1.26.0
pandas>=2.1.4
scipy>=1.11.4
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0

# Explainability
grad-cam>=1.4.8            # Grad-CAM heatmaps

# Backend API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
aiofiles>=23.2.1
httpx>=0.26.0

# Experiment Tracking
tensorboard>=2.15.1
wandb>=0.16.2

# Configuration
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.1
rich>=13.7.0
click>=8.1.7
psutil>=5.9.8

# Video Processing
imageio>=2.33.1
imageio-ffmpeg>=0.4.9

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
aiosqlite>=0.20.0
databases>=0.9.0
firebase-admin>=6.4.0